requirements.txt
# Core dependencies
langchain-anthropic>=0.1.0
langchain>=0.1.0

# Java AST parsing
javalang>=0.13.0
tree-sitter>=0.20.0
tree-sitter-java>=0.20.0

# Git operations
gitpython>=3.1.0

# Utilities
pyyaml>=6.0
rich>=13.0.0  # Better CLI output
click>=8.1.0  # CLI framework

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
mypy>=1.5.0
